+++ sklearn_genetic/__init__.py
@@ -7,6 +7,7 @@ from .callbacks import (
     LogbookSaver,
 )
 
+from .schedules import ExponentialAdapter, InverseAdapter, PotentialAdapter
 
 from ._version import __version__
 
@@ -17,5 +18,8 @@ __all__ = [
     "ConsecutiveStopping",
     "DeltaThreshold",
     "LogbookSaver",
+    "ExponentialAdapter",
+    "InverseAdapter",
+    "PotentialAdapter",
     "__version__",
 ]