+++ sklearn_genetic/schedules/schedulers.py
@@ -1,0 +1,106 @@
+import math
+from .base import BaseAdapter
+
+
+class ExponentialAdapter(BaseAdapter):
+    """
+    Adapts the initial value towards the end value using an exponential "decay" function
+
+    Parameters
+    ----------
+        initial_value : float,
+            Initial value to be adapted
+        end_value : float,
+            The final (asymptotic) value that the initial_value can take
+        adaptive_rate : float,
+            Controls how fast the initial_value approaches the end_value
+
+    Attributes
+    ----------
+        current_step : int,
+            The current number of iterations that the adapter has run
+
+        current_value : float,
+            The transformed initial_value after current_steps changes
+    """
+
+    def __init__(self, initial_value, end_value, adaptive_rate):
+        super().__init__(initial_value, end_value, adaptive_rate)
+
+    def step(self):
+        self.current_value = (
+            (self.initial_value - self.end_value)
+            * math.exp(-self.adaptive_rate * self.current_step)
+        ) + self.end_value
+        self.current_step += 1
+
+        return self.current_value
+
+
+class InverseAdapter(BaseAdapter):
+    """
+    Adapts the initial value towards the end value using a "decay" function of the for 1/x
+
+    Parameters
+    ----------
+        initial_value : float,
+            Initial value to be adapted
+        end_value : float,
+            The final (asymptotic) value that the initial_value can take
+        adaptive_rate : float,
+            Controls how fast the initial_value approaches the end_value
+
+    Attributes
+    ----------
+        current_step : int,
+            The current number of iterations that the adapter has run
+
+        current_value : float,
+            The transformed initial_value after current_steps changes
+    """
+
+    def __init__(self, initial_value, end_value, adaptive_rate):
+        super().__init__(initial_value, end_value, adaptive_rate)
+
+    def step(self):
+        self.current_value = self.end_value + (
+            (self.initial_value - self.end_value)
+            / (1 + self.adaptive_rate * self.current_step)
+        )
+        self.current_step += 1
+
+        return self.current_value
+
+
+class PotentialAdapter(BaseAdapter):
+    """
+    Adapts the initial value towards the end value using a potential "decay" function
+
+    Parameters
+    ----------
+        initial_value : float,
+            Initial value to be adapted
+        end_value : float,
+            The final (asymptotic) value that the initial_value can take
+        adaptive_rate : float,
+            Controls how fast the initial_value approaches the end_value
+
+    Attributes
+    ----------
+        current_step : int,
+            The current number of iterations that the adapter has run
+
+        current_value : float,
+            The transformed initial_value after current_steps changes
+    """
+
+    def __init__(self, initial_value, end_value, adaptive_rate):
+        super().__init__(initial_value, end_value, adaptive_rate)
+
+    def step(self):
+        self.current_value = self.end_value + (
+            self.initial_value - self.end_value
+        ) * math.pow((1 - self.adaptive_rate), self.current_step)
+        self.current_step += 1
+
+        return self.current_value