+++ sklearn_genetic/tests/test_genetic_search.py
@@ -22,6 +22,8 @@ from ..callbacks import (
     ProgressBar,
 )
 
+from ..schedules import ExponentialAdapter, InverseAdapter
+
 data = load_digits()
 label_names = data["target_names"]
 y = data["target"]
@@ -33,7 +35,7 @@ X_train, X_test, y_train, y_test = train_test_split(
 
 
 def test_expected_ga_results():
-    clf = SGDClassifier(loss="log", fit_intercept=True)
+    clf = SGDClassifier(loss="modified_huber", fit_intercept=True)
     generations = 6
     evolved_estimator = GASearchCV(
         clf,
@@ -150,7 +152,7 @@ def test_expected_ga_results():
     ],
 )
 def test_expected_algorithms_callbacks(algorithm, callback):
-    clf = SGDClassifier(loss="log", fit_intercept=True)
+    clf = SGDClassifier(loss="modified_huber", fit_intercept=True)
     generations = 8
     evolved_estimator = GASearchCV(
         clf,
@@ -287,7 +289,7 @@ def test_negative_criteria():
 
 
 def test_wrong_criteria():
-    clf = SGDClassifier(loss="log", fit_intercept=True)
+    clf = SGDClassifier(loss="modified_huber", fit_intercept=True)
     generations = 8
     with pytest.raises(Exception) as excinfo:
         evolved_estimator = GASearchCV(
@@ -338,7 +340,7 @@ def test_wrong_estimator():
 
 
 def test_wrong_get_item():
-    clf = SGDClassifier(loss="log", fit_intercept=True)
+    clf = SGDClassifier(loss="modified_huber", fit_intercept=True)
     generations = 8
     evolved_estimator = GASearchCV(
         clf,
@@ -392,7 +394,7 @@ def test_iterator():
 
 
 def test_wrong_algorithm():
-    clf = SGDClassifier(loss="log", fit_intercept=True)
+    clf = SGDClassifier(loss="modified_huber", fit_intercept=True)
     generations = 6
     evolved_estimator = GASearchCV(
         clf,
@@ -420,7 +422,7 @@ def test_wrong_algorithm():
 
 
 def test_no_param_grid():
-    clf = SGDClassifier(loss="log", fit_intercept=True)
+    clf = SGDClassifier(loss="modified_huber", fit_intercept=True)
     generations = 8
     with pytest.raises(Exception) as excinfo:
         evolved_estimator = GASearchCV(
@@ -493,7 +495,7 @@ def test_param_grid_one_param():
 
 
 def test_expected_ga_multimetric():
-    clf = SGDClassifier(loss="log", fit_intercept=True)
+    clf = SGDClassifier(loss="modified_huber", fit_intercept=True)
     scoring = {
         "accuracy": "accuracy",
         "balanced_accuracy": make_scorer(balanced_accuracy_score),
@@ -572,7 +574,7 @@ def test_expected_ga_multimetric():
 
 
 def test_expected_ga_callable_score():
-    clf = SGDClassifier(loss="log", fit_intercept=True)
+    clf = SGDClassifier(loss="modified_huber", fit_intercept=True)
     generations = 6
     scoring = make_scorer(accuracy_score)
     evolved_estimator = GASearchCV(
@@ -640,3 +642,85 @@ def test_expected_ga_callable_score():
     assert "std_fit_time" in cv_result_keys
     assert "mean_score_time" in cv_result_keys
     assert "params" in cv_result_keys
+
+
+def test_expected_ga_schedulers():
+    clf = SGDClassifier(loss="modified_huber", fit_intercept=True)
+    generations = 6
+
+    mutation_scheduler = ExponentialAdapter(
+        initial_value=0.6, adaptive_rate=0.01, end_value=0.2
+    )
+    crossover_scheduler = InverseAdapter(
+        initial_value=0.4, adaptive_rate=0.01, end_value=0.3
+    )
+
+    evolved_estimator = GASearchCV(
+        clf,
+        cv=3,
+        scoring="accuracy",
+        population_size=6,
+        generations=generations,
+        tournament_size=3,
+        mutation_probability=mutation_scheduler,
+        crossover_probability=crossover_scheduler,
+        elitism=False,
+        keep_top_k=4,
+        param_grid={
+            "l1_ratio": Continuous(0, 1),
+            "alpha": Continuous(1e-4, 1, distribution="log-uniform"),
+            "average": Categorical([True, False]),
+            "max_iter": Integer(700, 1000),
+        },
+        verbose=False,
+        algorithm="eaSimple",
+        n_jobs=-1,
+        return_train_score=True,
+    )
+
+    evolved_estimator.fit(X_train, y_train)
+
+    assert check_is_fitted(evolved_estimator) is None
+    assert "l1_ratio" in evolved_estimator.best_params_
+    assert "alpha" in evolved_estimator.best_params_
+    assert "average" in evolved_estimator.best_params_
+    assert len(evolved_estimator) == generations + 1  # +1 random initial population
+    assert len(evolved_estimator.predict(X_test)) == len(X_test)
+    assert evolved_estimator.score(X_train, y_train) >= 0
+    assert len(evolved_estimator.decision_function(X_test)) == len(X_test)
+    assert len(evolved_estimator.predict_proba(X_test)) == len(X_test)
+    assert len(evolved_estimator.predict_log_proba(X_test)) == len(X_test)
+    assert evolved_estimator.score(X_test, y_test) == accuracy_score(
+        y_test, evolved_estimator.predict(X_test)
+    )
+    assert bool(evolved_estimator.get_params())
+    assert len(evolved_estimator.hof) == evolved_estimator.keep_top_k
+    assert "gen" in evolved_estimator[0]
+    assert "fitness_max" in evolved_estimator[0]
+    assert "fitness" in evolved_estimator[0]
+    assert "fitness_std" in evolved_estimator[0]
+    assert "fitness_min" in evolved_estimator[0]
+
+    cv_results_ = evolved_estimator.cv_results_
+    cv_result_keys = set(cv_results_.keys())
+
+    assert "param_l1_ratio" in cv_result_keys
+    assert "param_alpha" in cv_result_keys
+    assert "param_average" in cv_result_keys
+    assert "split0_test_score" in cv_result_keys
+    assert "split1_test_score" in cv_result_keys
+    assert "split2_test_score" in cv_result_keys
+    assert "split0_train_score" in cv_result_keys
+    assert "split1_train_score" in cv_result_keys
+    assert "split2_train_score" in cv_result_keys
+    assert "mean_test_score" in cv_result_keys
+    assert "std_test_score" in cv_result_keys
+    assert "rank_test_score" in cv_result_keys
+    assert "mean_train_score" in cv_result_keys
+    assert "std_train_score" in cv_result_keys
+    assert "rank_train_score" in cv_result_keys
+    assert "std_fit_time" in cv_result_keys
+    assert "mean_score_time" in cv_result_keys
+    assert "params" in cv_result_keys
+
+    assert crossover_scheduler.current_value + mutation_scheduler.current_value <= 1